로그아웃에 비밀번호 입력창 추가. 현재 controller if문 문제로 값이 같으면 무조건 탈퇴 되는 문제가 있음. 수정 예정.

+++ TheRecipe/src/main/java/com/increpas/therecipe/controller/MemberController.java
@@ -124,56 +124,71 @@ public class MemberController {
 
 	}
 
-	
-	@RequestMapping(value="/idcheck.do", method = RequestMethod.GET)
-	public String IdCheck(){
-		
+	@RequestMapping(value = "/idcheck.do", method = RequestMethod.GET)
+	public String IdCheck() {
+
 		return "idcheck";
 	}
-	
+
 	@RequestMapping(value = "/idchecksuccess.do", method = RequestMethod.POST)
 	public String IdCheck(Model model, HttpServletRequest request) {
-		
-		String checkId = request.getParameter("checkid");
 
+		String checkId = request.getParameter("checkid");
 
 		int userCnt = Integer.parseInt(memberService.checkId(checkId));
 
 		String msg = "";
-		if( userCnt >= 1){
+		if (userCnt >= 1) {
 			msg = "사용중인 아이디 입니다.";
-		}else {
+		} else {
 			msg = "사용하실 수 있는 아이디 입니다.";
 		}
 		model.addAttribute("msg", msg);
-		
+
 		return "join";
 	}
-
 
 	/**
 	 * 회원탈퇴
 	 * 
-	 * @param vo MemberVO vo
-	 * @param errors Errors errors
-	 * @param request HttpServletRequest
-	 * @param session HttpSession session
+	 * @param vo
+	 *            MemberVO vo
+	 * @param errors
+	 *            Errors errors
+	 * @param request
+	 *            HttpServletRequest
+	 * @param session
+	 *            HttpSession session
 	 * @return 회원탈퇴 성공 시 홈 화면으로 이동
 	 */
-	@RequestMapping(value = "/delete_Id.do", method = {RequestMethod.GET,RequestMethod.POST})
-	public String deleteIdViewGet(@Valid MemberVO vo, Errors errors, HttpServletRequest request, HttpSession session) {
+
+	@RequestMapping(value = "/delete_Id.do", method = RequestMethod.GET)
+	public String deleteIdViewGet() {
+		return "deleteRePw";
+	}
+
+	// @ModelAttribute("deleteGo") @Valid
+	@RequestMapping(value = "/delete_Id_Go.do", method = RequestMethod.POST)
+	public String deleteIdViewPost(@Valid MemberVO vo, Errors errors, HttpServletRequest request, HttpSession session) {
 		if (errors.hasErrors()) {
 			logger.info("회원탈퇴 유효성체크 오류발생");
 		}
-		
-		String m_userid = (String) session.getAttribute("m_userid");
-		System.out.println(m_userid);
-		
-		memberService.deleteId(m_userid);
-		
-		session.invalidate();
 
-		return "home";
+		String repw = request.getParameter("m_repw");
+		String m_pw = request.getParameter("m_pw");
+
+		if (!repw.equals(m_pw)) {
+
+			return "deleteRePw";
+
+		} else {
+
+			String m_userid = (String) session.getAttribute("m_userid");
+
+			memberService.deleteId(m_userid);
+			session.invalidate();
+			return "home";
+		}
 
 	}
